Refactor year-end report classes and request types
Updated properties in `YearEndProfitSharingReportResponse` and `ProfitShareTotal` to use `long` for employee counts. Changed method signatures in `IProfitSharingSummaryReportService` and related classes to accept `ProfitYearRequest` instead of `FrozenProfitYearRequest`. Streamlined calculations in `ProfitSharingSummaryReportService` and removed unnecessary using directives in endpoint files.

+++ src/services/src/Demoulas.ProfitSharing.Data/Entities/Virtual/ProfitShareTotal.cs
@@ -10,7 +10,7 @@ public sealed class ProfitShareTotal
     public decimal TerminatedHoursTotal { get; set; }
     public decimal TerminatedPointsTotal { get; set; }
     public decimal TerminatedBalanceTotal { get; set; }
-    public int NumberOfEmployees { get; set; }
+    public long NumberOfEmployees { get; set; }
     public int NumberOfNewEmployees { get; set; }
     public int NumberOfEmployeesUnder21 { get; set; }
 }
